Preparing for Context Aware Nodes (WIP)
