+++ src/nodeserver/wrapper/nodes/node/node_utils.py
@@ -3,10 +3,10 @@ from collections import deque
 from nodeserver.wrapper.nodes.data.node_data_types import SuperSlotTypes
 from nodeserver.wrapper.nodes.node.base_nodes import _ParsedNode, NodeMirror
 
-class NodeUtils:
+class NodeMirrorUtils:
     @staticmethod
     def get_node_execution_order(nodes: list[_ParsedNode]) -> list[NodeMirror]:
-        execution_order = NodeUtils._get_node_execution_order([node._mirror for node in nodes])
+        execution_order = NodeMirrorUtils._get_node_execution_order([node._mirror for node in nodes])
         return execution_order
 
     # Obrigado Gemini por fazer esse algoritmo aí, tamo junto
